requirements.txt
Flask
python-dotenv
pandas
python-docx
PyMuPDF
openpyxl
cohere
chromadb
requests
